Fixed circular reference
